Fixed Compliance rules

+++ src/main/java/com/thjug/bgile/facade/CardFacade.java
@@ -1,6 +1,14 @@
 /*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
+ * Attribution
+ * CC BY
+ * This license lets others distribute, remix, tweak,
+ * and build upon your work, even commercially,
+ * as long as they credit you for the original creation.
+ * This is the most accommodating of licenses offered.
+ * Recommended for maximum dissemination and use of licensed materials.
+ *
+ * http://creativecommons.org/licenses/by/3.0/
+ * http://creativecommons.org/licenses/by/3.0/legalcode
  */
 package com.thjug.bgile.facade;
 
@@ -116,7 +124,7 @@ public class CardFacade {
 		card.setOwner((tostate == State.Plan.getId()) ? null : new Account(accountid));
 
 		// FIXME: Should refactor to one method.
-		if (fromstate == tostate) {
+		if (fromstate.intValue() == tostate.intValue()) {
 			final DashboardColumn columnstart = columns.get(fromstate);
 			final StringBuilder result = new StringBuilder();
 			for (final String s : columnstart.getWidgets()) {
